refactor: split `sponsored-call.helper`
